01_Application_Framework: 01_Introduction.md
This formats the entire page to at most 80 characters wide pagraphs.

Mentions that with 'needlefish' we started using gRPC and that we've
started to deprecate D-Bus, including the fact that we no longer spawn
directly applications, that we use system units for listing
applications. Most the of page is kept intact with those minor changes
being added.

Bug-AGL: SPEC-4700
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Id19e6cea90fdbe9a6bd4eee851f83b9d0f800db5
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/documentation/+/28521
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>